Refactor User page and NavBar component
My original structure had the Outlet component inside the NavBar
component. This is not desirable as the NavBar was no longer an
independent component, and was instead doing two things.

I refactored the NavBar so that its only purpose is to contain
the NavLinks, then refactored the '/user' route to instead be
a nested route with '/user' route rendering the UserPageLayout which
is a component that has the NavBar and the Outlet component. This
results in a cleaner representation of the layout of the pages that
include the NavBar as the layout is controlled soley by the
UserPageLayout component.