Refactor code to use SessionExercises classes
The structure of the sessionExercises state object is not basic. As
a result, when updating the state there was repeated code to perform
alterations to the object that spanned 10-20 lines of code.

This would be hard to maintain and read. So in order to fix this I
created a SessionExercises, Exercise and Set class with appropreate
methods so I could reduce this repeated code and make it far more
readable.