Refactor addUserToDb functions into one function
After refactoring the addEmailUserToDb function it became apparent that
the process of adding user data to the database did not depend on
whether they were signing up with Google or by using an email and
password, the difference lied in what object was being passed to the
function.

Thus the addEmailUserToDb was refactored to addUserToDb to handle
either signup method, and the affected code and tests were updated
accordingly.