Fix bug caused by invalid parameter possibility.
The constructor could potentially receive an empty object. We needed to
ensure the received value is not an empty object, as several lines
down we attempt to call Object.keys on this objects _exercises property,
which would be undefined if it did not exist (as in the case of an
empty object).