Refactor components to use a global context state
When adding additional components to the Dashboard page (all of which
used the users exercise history data) it became apparent that it would
be cleaner for there to be a single call for all the data when the
Dashboard component mounts, then share this with all children components.
Originally each child component would fetch sections of the data, but
this becomes inefficient and hard to maintain.