Refactor shared logic of graph component
After writing the VolumePerSessionGraph component I realised there
was alot of logic I wanted to use when graphing other data. In particular
the logic for the control buttons to move forward and backward through
the data, as well as showing all the data when the 'All' button was
clicked.

To reuse this functionality I wrote a more general component called
GraphWithControlBtns which then accepted a function as a prop
that parses the user history data and returns the data in the required
form, based on what you want to display in the graph.