Update routing tests
There were several additional components added to several pages
after the routing tests were written. After these were added these
components would throw uncaught errors in the routing tests because
they would be attempting to access context that had not been set up
in the original tests, or they were making async calls to the database
for user data.

Given the routing tests were written to test that the routing works
correctly, the providers and components that made these calls were mocked
in order to resolve these uncaught errors in the tests. This way
we continue to test solely with the intent of ensuring the routing works,
and these subcomponents can then be tested in isolation.