Fix bug re extra fetchNextResults calls
If the user is on the History page and moves to a different page
or logs out the fetchNextResults function will make a call. This call
is not needed, and in the case of logging out it throws an error as
the user is not logged in and so the async firestore call throws an error.

To resolve this issue the initial call to get the first page of data
is moved to the History component instead of the custom hook.