Refactor sign in method to use redirect
The signInWithPopup method would throw an error warning that the
Cross-Origin-Opener-Policy policy would block the window.closed call.
While the functionality still worked, to avoid this policy potentially
breaking the google sign in functionality I considered using signing
in with redirect.

After testing this functionality I chose to make the change perminent
as the UX seems nicer and it would work much better on mobile devices.