Refactor all media queries to the old syntax
I had used the newer syntax for media queries
(eg @media and (width <= 500px)) which I understood to be addopted on
most modern browsers. This worked when serving locally on my machine.

However after deploying the app and testing on my phone the media queries
werent applying. So I tested changing one of the pages media queries
to the older syntax and the site became responsive.