Fix useMediaQuery hook usage
The new media query syntax was used for this hook, however when
testing on a mobile device the responsiveness failed.

Additionally the boolean value is opposite to what I expected
- ie to check if it is a small screen I need to get the boolean
value of 'max-width: 900px' which is true if the screen is small, as
opposed to 'min-width: 900px' which would be false for a small screen.

As such all uses of the useMediaQuery hook were updated.